feat(rename): :twisted_rightwards_arrows: add typing rename demo
